Fix : Scripts : Correct Highscore updation after collision
The high-score is recorded by saving the player-data in an external file,
which should be loaded before the mainScreen is loaded. The problem with
the current implementation was that the class MoveByTouch didn't load the
data upon startup, and hence a value of 0 was assigned to the `highScore`
variable everytime the object was created, thus while updating the
highScore after the game ended, the implementation checked if the
currentScore was greater than the current highScore or not, thus
assigning it the value of currentScore everytime (since highScore=0).

Thus, modify the class to load high-score from the external file in
the `Start()` method of MoveByTouch.

Signed-off-by: Suraj Upadhyay <usuraj35@gmail.com>